requirements.txt
# Core
aiohttp==3.8.4
asyncio==3.4.3
pyyaml==6.0
python-dotenv==1.0.0

# GUI
pygame==2.5.2

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Networking
asyncssh==2.13.2
aiofiles==23.2.1

# Security
cryptography==41.0.7
bcrypt==4.1.2

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Database (optional)
sqlalchemy==2.0.23
aiosqlite==0.19.0
elasticsearch==8.11.0

# Dashboard
jinja2==3.1.2
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1